requirements.txt
flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
mysql-connector-python==8.0.33 